requirements.txt
requests
gspread
gspread-dataframe
google-auth
pandas
pytz
python-dotenv
openpyxl
